Clean up database test fixtures
In order to work seamlessly with a local database, we also need to use the same database engine that the connection was created from. Otherwise we run the risk of e.g. connecting to the prod database and running a bunch of delete statements.
